Add CHD copy/re-compression, uncompressed CHDs, and delta creation to the encoder
Phase 4 of the ProposedFixes plan, completing full chdman feature parity for
CHD creation:

- ChdEncoder.Copy: re-compress any V1-V5 CHD (child sources via
  ChdEncodeOptions.SourceParentPath), cloning all metadata; shared EncodeCore
  pipeline refactor unifies EncodeRaw/EncodeCd/Copy
- ChdEncoder.EncodeUncompressed (-c none): V5 raw map, hunk-aligned raw data,
  zero-hunk skip, no SHA-1 - byte-for-byte identical to chdman createraw -c none
- ChdCodecs.CreateAll accepts [None] as an empty codec list
- CLI: --copy command (-c/-t/-ip/-op/-v); -c none on --create/--createcd/--copy
- FLAC encoder: FlacSubframe gains a window parameter (call sites updated)
- Tests: 350 encoder + 563 lib tests on net8/9/10 (23 new: ChdCopyTests,
  UncompressedEncodeTests), chdman verify/extractraw byte-identical validation
- Docs: READMEs and docs/ updated (copy, -c none, delta children)

+++ CHDSharpEncoderTest/MapCompressorTests.cs
@@ -97,7 +97,7 @@ public class MapCompressorTests
             entries[i] = new MapEntry { Compression = MapEntry.CompressionType0, CompLength = 50, Offset = (ulong)(124 + i * 60), Crc16 = 0 };
         }
 
-        byte[] compressed = MapCompressor.Compress(entries, (uint)count, 4096, 512);
+        byte[] compressed = MapCompressor.Compress(entries, count, 4096, 512);
 
         // The compressed map should be much smaller than 100 × 12 = 1200 bytes
         Assert.True(compressed.Length < 1200);
